dom/xml/tests/mochitest/test_bug399502.xhtml

<?xml version="1.0"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=399502
-->
<head>
  <title>Test for Bug 399502</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=399502">Mozilla Bug 399502</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

/** Test for Bug 399502 **/

var xmldoc;
function runTest() {
  xmldoc = document.implementation.createDocument("", "", null);
  xmldoc.onload = function(evt) {
                    ok(true, "Onload handler didn't get called!");
                    SimpleTest.finish();
                  };
  xmldoc.load("test_bug399502.xhtml");
}
SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);


]]>
</script>
</pre>
</body>
</html>
